% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllPlantsby2020.R
\docType{data}
\name{AllPlantsby2020}
\alias{AllPlantsby2020}
\title{Bon Thomas A All Plants by 2020}
\format{
A data frame with 77 rows and 3 variables:
\describe{
  \item{Tag ID}{Plant Tag identification}
  \item{x-axis}{x-axis position}
  \item{y-axis}{y-axis position}
}
}
\source{
Dr Megan O'Shea: Spiny Rice Flowers Project
}
\usage{
AllPlantsby2020
}
\description{
Existing plants plus recruits over 2017 to 2020
 The variables are as follows:
}
\keyword{datasets}
