% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BonThomasARecruit2021.R
\docType{data}
\name{BonThomasARecruit2021}
\alias{BonThomasARecruit2021}
\title{BonThomasARecruit2021}
\format{
A data frame with 77 rows and 3 variables:
\describe{
  \item{Tag ID}{Plant Tag identification}
  \item{x-axis}{x-axis position}
  \item{y-axis}{y-axis position}
}
}
\source{
Dr Megan O'Shea: Spiny Rice Flowers Project
}
\usage{
BonThomasARecruit2021
}
\description{
Plants that are recruits in 2021
 The variables are as follows:
}
\keyword{datasets}
