% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calctot.R
\name{calctot}
\alias{calctot}
\title{Calculate Total}
\usage{
calctot(ss, probs, incprobs, qx)
}
\arguments{
\item{ss}{Sample Size}

\item{probs}{Quadrat Probabilities}

\item{incprobs}{Inclusion probabilities}

\item{qx}{Quadrat Counts}
}
\value{
Estimated Total Number
}
\description{
Calculate Total
}
\examples{
\dontrun{
ff <- ks::kde.boundary(AllPlantsby2020[,2:3],
                    xmin=c(0,0),xmax=c(35,15))
probs1 <- quadratprobs(ff, quadratstart=c(0,0),quadratsize=5)
results <- table(replicate(100,calctot(15,
probs=probs1,
incprobs=BonThomasAIncProbsMat[15,],
qx=QuadratChange5_2021)))
results
Hmisc::wtd.mean(as.numeric(names(results)), results)
sqrt(Hmisc::wtd.var(as.numeric(names(results)), results))
}
}
