% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Change.R
\name{Change}
\alias{Change}
\title{Quadrat counts of changes in a particular year}
\usage{
Change(site, year, quadratsize = 5, quadratstart = c(0, 0))
}
\arguments{
\item{site}{Sampling site}

\item{year}{Year}

\item{quadratsize}{Size of quadrats (m)}

\item{quadratstart}{Lower left starting position of quadrats}
}
\value{
A contingency table containing the change (recruits - died) in each quadrat
}
\description{
Quadrat counts of changes in a particular year
}
\examples{
Change(BonThomasA, 2021, quadratsize=5)

Change(BonThomasA, 2021, quadratsize=4, quadratstart=c(2,1))

}
