% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DentonAveB.R
\docType{data}
\name{DentonAveB}
\alias{DentonAveB}
\title{Denton Ave B Data 2017-2021}
\format{
A data frame with 150 rows and 8 variables:
\describe{
  \item{Tag ID}{Plant Tag identification}
  \item{Status 2017}{Status as of 2017: Categorical variable with levels "Existing", "Germinant", and "NA" (Missing)}
  \item{Status 2018}{Status as of 2018: Categorical variable with levels "Alive","Dead","Recruit","Germinant", and "NA" (Missing)}
  \item{Status 2019}{Status as of 2019: Categorical variable with levels "Alive","Dead","Recruit","Germinant", and "NA" (Missing)}
  \item{Status 2020}{Status as of 2020: Categorical variable with levels "Alive","Dead","Recruit","Germinant", and "NA" (Missing)}
  \item{Status 2021}{Status as of 2021: Categorical variable with levels "Alive","Dead","Recruit", and "Germinant"}
  \item{x-axis}{x-axis position}
  \item{y-axis}{y-axis position}
}
}
\source{
Dr Megan O'Shea: Spiny Rice Flowers Project
}
\usage{
DentonAveB
}
\description{
Longitudinal data for Denton Ave B
}
\details{
The dimensions of this grid are:
 Length = 36m., Width = 20m., Area = 720 sq. m.
}
\keyword{datasets}
