% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Died.R
\name{Died}
\alias{Died}
\title{Location of plants that have died in a particular year}
\usage{
Died(site, year)
}
\arguments{
\item{year}{}
}
\value{
A data-frame of plants that have died with Tag ID, x-axis, and y-axis
}
\description{
Location of plants that have died in a particular year
}
\examples{
Died(BonThomasA, 2021)
}
