% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_tot.R
\name{est_tot}
\alias{est_tot}
\title{Estimated Total of Plants}
\usage{
est_tot(nsamp = 6, mm = mm5, qx = qx5, rr = rr5)
}
\arguments{
\item{nsamp}{Number of Samples}

\item{mm}{Not sure what it is}

\item{qx}{Quadrat Counts}

\item{rr}{Output from rrgen}
}
\value{
Coefficient of variation
}
\description{
Estimated Total of Plants
}
