% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMap.R
\name{HeatMap}
\alias{HeatMap}
\title{Plot Heat Map}
\usage{
HeatMap(
  xy,
  weights = rep(1, nrow(xy)),
  xrange = c(0, 35),
  yrange = c(0, 15),
  gridsep = 5,
  title = "Heat Map of Location"
)
}
\arguments{
\item{xy}{Data Frame to use}

\item{xrange}{Limits of the x-variable}

\item{yrange}{Limits of the y-variable}

\item{gridsep}{Separation of dotted lines}
}
\value{
a ggplot2 object
}
\description{
Plot Heat Map
}
\examples{
HeatMap(Site29[,2:3], weights=Site29$NUMPimelea,xrange=c(11,25), yrange=c(0,11), gridsep=5, title="Heat Map of Site 29")
}
