% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclusionprobs.R
\name{inclusionprobs}
\alias{inclusionprobs}
\title{Title}
\usage{
inclusionprobs(quadratprobs, nsim = 10000)
}
\arguments{
\item{quadratprobs}{quadrat probabilies}

\item{nsim}{Number of simulations}
}
\value{
Vector of inclusion probabilities
}
\description{
Title
}
\examples{

inclusionprobs(c(0.5,0.15,0.3, 0.05),nsim=100000)
}
