% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PimeleaE.R
\docType{data}
\name{PimeleaE}
\alias{PimeleaE}
\title{Longitudinal data for Pimelea E}
\format{
A data frame with 15 rows and 8 variables:
\describe{
  \item{Tag ID}{Plant Tag identification}
  \item{Status 2017}{Status as of 2017}
  \item{Status 2018}{Status as of 2018}
  \item{Status 2019}{Status as of 2019}
  \item{Status 2020}{Status as of 2020}
  \item{Status 2021}{Status as of 2021}
  \item{x-axis}{x-axis position}
  \item{y-axis}{y-axis position}
}
}
\source{
Dr Megan O'Shea: Spiny Rice Flowers Project
}
\usage{
PimeleaE
}
\description{
The variables are as follows:
}
\keyword{datasets}
