% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadprob.R
\name{quadprob}
\alias{quadprob}
\title{calculate}
\usage{
quadprob(xs = xx, ys = yy, xbr = seq(0, 35, 5), ybr = seq(0, 15, 5))
}
\arguments{
\item{xs}{x-coordinates of the data}

\item{ys}{y-coordinates of the data}

\item{xbr}{quadrat x-breaks}

\item{ybr}{quadrat y-breaks}
}
\value{
matrix of quadrat probabilitiesx
}
\description{
calculate
}
\examples{
\dontrun{
x <- AllPlantsby2020[,"x-axis"]
y <- AllPlantsby2020[,"y-axis"]
quadprob(x,y)
}

}
