% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BonThomasADied2021.R
\docType{data}
\name{QuadratChange5_2021}
\alias{QuadratChange5_2021}
\title{QuadratChange5_2021}
\format{
A matrix with 3 rows and 7 columns:
}
\source{
Dr Megan O'Shea: Spiny Rice Flowers Project
}
\usage{
QuadratChange5_2021
}
\description{
Quadrat counts of chnges between 2020 and 2021
 The variables are as follows:
}
\keyword{datasets}
