% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site22.R
\docType{data}
\name{Site22}
\alias{Site22}
\title{Site22}
\format{
A data frame with 262 rows and 4 variables:
\describe{
  \item{id_2}{1m by 1m quadrat id}
  \item{x-axis}{mid-point of quadrat x-axis position}
  \item{y-axis}{mid-point of quadrat y-axis position}
  \item{NUMPimela}{Count of number of plants in the quadrat}
}
}
\source{
Dr Megan O'Shea: Spiny Rice Flowers Project
}
\usage{
Site22
}
\description{
Counts of Plants at Site 22 Hands Road
}
\keyword{datasets}
