% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site29.R
\docType{data}
\name{Site29}
\alias{Site29}
\title{Site29}
\format{
A data frame with 12 rows and 4 variables:
\describe{
  \item{id_2}{1m by 1m quadrat id}
  \item{x}{mid-point of quadrat x-axis position}
  \item{y}{mid-point of quadrat y-axis position}
  \item{NUMPimelea}{Count of number of plants in the quadrat}
}
}
\source{
Dr Megan O'Shea: Spiny Rice Flowers Project
}
\usage{
Site29
}
\description{
Counts of Plants at Site 29 Jasper Rd Patch220
}
\keyword{datasets}
