% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site45.R
\docType{data}
\name{Site45}
\alias{Site45}
\title{Site45}
\format{
A data frame with 573 rows and 4 variables:
\describe{
  \item{id_2}{1m by 1m quadrat id}
  \item{x}{mid-point of quadrat x-axis position}
  \item{y}{mid-point of quadrat y-axis position}
  \item{NUMPimela}{Count of number of plants in the quadrat}
}
}
\source{
Dr Megan O'Shea: Spiny Rice Flowers Project
}
\usage{
Site45
}
\description{
Counts of Plants at Site 45 Phylands
}
\keyword{datasets}
