% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumberAlive.R
\name{NumberAlive}
\alias{NumberAlive}
\title{Number of Alive Plants at a Site in a particular year}
\usage{
NumberAlive(Site, year)
}
\arguments{
\item{Site}{Site}

\item{year}{Year}
}
\value{
The number of alive plants
}
\description{
Number of Alive Plants at a Site in a particular year
}
\examples{
NumberAlive(BonThomasA, 2020)
}
