% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratprobs.R
\name{quadratprobs}
\alias{quadratprobs}
\title{Title Calculate Quadrat Probability}
\usage{
quadratprobs(
  plantlocations,
  quadratstart = c(0, 0),
  quadratsize = 5,
  xmin = c(0, 0),
  xmax = c(35, 15)
)
}
\arguments{
\item{plantlocations}{x and y coordinates}

\item{quadratstart}{lower left co-ordinates ofthe first quadrat}

\item{quadratsize}{width (and length) of the quadrats}

\item{xmin}{vector of grid minima}

\item{xmax}{vector of grid maxima}
}
\value{
matrix of quadrat probabilities
}
\description{
Title Calculate Quadrat Probability
}
\examples{

quadratprobs(AllPlantsby2020[,2:3], quadratstart=c(0,0),quadratsize=5,xmin=c(0,0),xmax=c(35,15))

}
