% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recruits.R
\name{Recruits}
\alias{Recruits}
\title{Location of plants that are recruits in a particular year}
\usage{
Recruits(site, year)
}
\arguments{
\item{site}{Site}

\item{year}{Year}
}
\value{
A data-frame of plants that are recruits with Tag ID, x-axis, and y-axis
}
\description{
Location of plants that are recruits in a particular year
}
\examples{
Recruits(BonThomasA, 2021)
}
