% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampres.R
\name{sampres}
\alias{sampres}
\title{Quadrat Probabilities}
\usage{
sampres(
  quadratsize,
  xmax = 35,
  ymax = 15,
  Allplants = AllPlantsby2020[, 2:3],
  xxmin = c(0, 0),
  xxmax = c(35, 15)
)
}
\arguments{
\item{quadratsize}{Size of quadrat ib m.}

\item{xmax}{Maximum x co-ordinate}

\item{ymax}{Maximum y co-ordinate}

\item{Allplants}{Locations of Plants}

\item{xxmin}{Minimum of x and y}

\item{xxmax}{Maximum of x and y}
}
\value{
Estimated quadrat  probability
}
\description{
Quadrat Probabilities
}
\examples{
sampres(5, Allplants = AllPlantsby2020[,c(2,3)])

}
